This commit changes the behavior of the matching selection in the search box's drop-down & autofill content.  When the first match would be the defaultText (e.g. "Search Rows..."), the match is updated to skip the default.
This happens upon keyReleased and the selection and autofill text as
updated to match.  It will still match the "Search Rows..." string if
there is no other valid match.

Now that I think about it, this means that we could be matching
wildcards in this code to allow the user to see a preview of what they
would get...  Another commit perhaps...