ResetDefault for ModelViews, work on configNode
A major problem when loading a new model at the moment is that model views
are not properly reset. This leads to all sorts of inconsistencies and
goes as far as interfering with remembering and setting headers upon
opening a new file which makes this problem a part of this issue.

This commit introduces some rudimentary approaches to reset important
member variables to a default state upon loading a new model and then
synchronizing their state from a Preferences node with a more
straight-forward interface ConfigNodePersistent. THis has also revealed
tons of design problems such as unnecessary configNodes in classes which
do not use them, bad exception handling, and more. This issue has turned
out to be extremely hairy and deep.