Handling old file when loading a new one
Made sure that the old Preferences from a file are taken over to the newly
loaded one in case they are related. This might be the case if a file gets
clustered - the preferences will be transferred. Or in case a file has
previously been loaded. Then the preferences will be imported.

For some reason changing 2 lines of code in LinkedViewApp and TreeViewApp
messed with the UIManager using the system L&F... investigating...