Tied the ExportHandler to the export interface.
After implementing this, I realized that there was not a way in the
requirements/design/mockup in BB issue #292 which allowed the user to
enter an output file name, location, or a way to create a new folder.
It also doesn't have a way to disable the papersize menu when an
incompatible format (e.g. png) is selected.