Can now completely hide/show trees and no more whizzing tree background exceptions.
With this commit, I have fixed 2 issues that have each been marked as
"resolved" even though they were not completely resolved.

The ScaleEq being passed in to the whizzing tree background code was
null after a new file was opened.  That is not checked.

The minimum size of the tree and label panes has to be 0 in order to
allow the divider to be dragged to the extreme positions.