Changed "showWarning" to "LogBuffer.println" to finish up issue #344, whose prior resolution was missing that tiny fix.
There may be other more serious causes of data loading interruption that
could possibly be discerned to necessitate a warning thrown to the user,
but I can't think of how to catch such a scenario.  Besides, if there is
such a scenario, I don't think that the location of the generation of
this message is where you would catch it.