Limited thumb label size for color chooser
The labels can not surpass a certain length until round and/ or switch to
scientif notation happens. An exception case for problematic gradient
drawing (non-increasing fractions) has also been added alongside a hint
for users to fix the problem. The issue will not crash and block the UI
anymore but provides the user with an opportunity to fix problems right
away.

Resolves: #366

See also: -