Resolved issue 341.  Label index is now checked against the size of the array before accessing it.
When the user opened a file with fewer labels than the previously opened
file and the longest label in the previous file was at an index out of
range from the previous file, a null pointer exception was thrown.  I
added a check of the index and the string at that index to make sure
things haven't changed before relying on that index.

Funnily enough, there was a TODO in the code to implement this check and
re-initialize the appropriate values.