Removed right-click to deselect from labelviews
Right-clicking to deselect is not a useful option anymore because right-clicking already brings up the context menu via the overriden mouseClicked() function in the LabelView children. The original mouseClicked function is ALSO called and whenever a left mouseclick does not happen (= some other click event), all indeces would be deselected. This was problematic when right-clicking to copy selected labels.