Added some exception catching and error dialogs.
One solution for the test 6kx6k matrix was to reduce the minimum tile
dimension size to 3x3, but I haven't changed that here yet in order to
test the exception catching.

I'm committing now so that I can go back to master and see if I can get
the same exception to happen (and figure out why I'm not getting it
right now with this code).