Swapped FileSet to String identifier for import
FileSet objects seem to self-update themselves periodically with
information stored in the active Preferences node 'FileSet'. Due to this, an old
FileSet object can not be used as a base to import old data. The relevant
identifiers (file root name, file extension) must be supplied directly so
the old Preferences node from which to import data can be found.