Added the ability to enable/disable and update the default selection of the export aspects offered in the export dialog.
So now, if the data can be exported as 1:1, but ends up being too big to
export with the aspect ratio set to "as seen on screen", that option is
disabled with a tooltip explaining why.

It is dynamically updated in this manner:

The selection of a format triggers an update of the available &
default-selected regions, which in turn, triggers an update of the
available & default-selected aspects.

The selection of a region triggers an update of the available &
default-selected aspects.

I tested this functionality using the large_6kx6k.txt file and changing
the minTileDim and the associated height & width (in ExportHandler) each
to 6.  This happens to result in a total image size that is too big when
as-seen-on-screen is selected and not too big at 1:1.  The app window on
my screen is set as small as it can go.  The effective aspect ratio
on-screen turns out to be 1.588.

There are still some debug prints that need cleaned up.  Plus, the
ability to disable image formats to be selected from the format select
list has not yet been implemented.