organized GUI component reset and state restoration
The state of the main components belonging to DendroView is explicitly
being reset to default before generating a new layout and after loading
the model. When all GUI components are set up, the Preferences settings
are restored as normal.

MapContainer.setIndexRange() now sets the first visible tile index as well
as the first visible label as minIndex. This is a bug fix because setSubImage() in
MatrixView never had the first visible tile index available upon initial
load after implementation of the resetDefaults() methods. An exception was
thrown.