A few small changes.
I removed a parameter from the old paint method which is set to null
everywhere it's called from (geneOrder).

I removed a wrapper paint method made to export the entire matrix
without supplying start indexes in each dimension.  You still had to
supply the size of each dimension, which could probably have also been
eliminated, but I didn't think the wrapper really solved a problem, so I
took it out and adjusted the single call to it to use the same method
that the other matrix drawing calls made.