Resolved issue #369 by setting the default close operation to "DO NOTHING".  Also commented out a call to System.exit(0) because it seemed to be quitting successfully and in a timely manner without it.  See issue #74 for more info.
Since quitting the app has a confirmation window upon using the close
button, we should dispose of the window conditionally based on the
response given from the dialog.  Since the default action had been set
to dispose on close, it was quitting the app and closing the window
regardless of the option selected.

Perhaps in the future, we can leave the app running, but also allow the
window to be closed, but for now, that's not possible.