Created the CustomDetailsConfirmDialog class which is used by TVController to create a dialog for resetting the preferences.  The dialog has a "Details" checkbox to reveal extra information on what happens when you reset preferences.
This version works, however, it has useless code that needs to be pulled
AND it's totally done the wrong way - using setMaximumSize, setSize, and
sets static values for the dialog size.  As yet, I've been unable to
figure out a way to force the text area to not make the dialog widen to
make the details string soft-wrap at a reasonable width.