Changed the showDialog method to not use static sizes for its elements and instead inserted line wraps into the detailed message so that a dynamically sized dialog window could be generated.
The determination of the line wrap size is a little kludgey because the
font size in the scrollpane appears smaller than the font size of the
summary message in the jlabel above it, so I just increased the number
of characters allowed on a line compared to the summary message by 10%,
which appears good.  It doesn't really matter too much because the
dialog window will expand if it ends up too big.

There's 1 issue that could arise, and that's if the summary message is
too long and the width of the window is statically set to 500, as is the
case in the code.  This will cause the calculation of the details pane
to be too big.  This is not a serious issue though since I don't use it
for any other strings, so we can address this if it becomes an issue.