Added new JTextArea for warning
Instead of changing the label for the input field upon an error, another
JTextArea was added beneath the input field in EditThumbDialog. This one
will display (and if necessary wrap) the error message to prevent the UI
from getting messed up.

Also, the creation code for text-only buttons was moved to GUIFactory and
made public static. This way we can reuse the code in case we need it in
the future.

Resolves: -
See also: -