I believe I have fixed the issue where sometimes zooming out would happen all on 1 axis and then all on another axis.
There was an error in the calculation of the number of cells to zoom
based on the zoom fraction sent in.  The correct code was there, but it
was commented out.  The code it had been replaced with was only correct
for zoom fractions less than or equal to 1 and even then, it was zooming
too much.  The zoom fraction is allowed to be larger than 1 when zooming
out.  I just had to multiply the number of visible cells by the target
zoom fraction.