Added versioning for Preferences API files
In order to distinguish updated versions of Preferences nodes, the
Preferences file stored on the system now has a version tag at its root.
Routines in ViewFrame.java, where the global configNode is located as a
member variable, handle the versions of Preferences and are (bare-bones)
set up to deal with version updates.

Resolves: #406
See also: #407