Removed useless code.
I realized that tracking justification and labal types selected for
display was not necessary in LabelView because the only time those ever
change is when the user does it in the label dialog.  It's the
resetSecondaryScroll call in the listeners that reset the label views,
so I took out the code in LabelView that tracks the selected label types
to display.