Resolves issue #415.  The zoom selection button now activates immediately upon selection change even if focus is on a search box.
Had to move the call to setAdaptiveButtonStatus to the top of the method
for requesting focus on the zoom selection button because it was
returning immediately if focus is on a search box.  As I recall, that
return was added so that a user can make multiple searches without
having to tab back to the search box each time.