Resolved steps 4 (and previously 7, which I forgot to mention in the previous commits) from BB issue 394.  The user is now brought back to the load dialog upon load failure so they can see if there is anything in the load interface they can change to fix the problem.
In TVController, if the load failed, it simply opens the import dialog
again using the same file (which was stored in a new datamember called
"loadingFile" during the initial load attempt).

Remaining issues:

Step 1: An error should be caught before clicking continue in the import
interface.  This is somewhat optional/ideal, so I don't think I'll
implement it.

I could not get the open last file button to become the default button
upon returning to the welcome screen.

Finally, in working on this, I realized that if the failed load came
from opening a file after a file had already been successfully loaded,
the user should be sent back to their previously loaded file instead of
to the welcome screen.  I think that this means that the PROGRESS_VIEW
should be removed from the main window and a separate progress window
should be created.  That would allow the previous data to remain loaded
so that upon failure, the user can go back to the data they were working
with before the failed load.