This resolves my issue 3 from the pull request review.  Highlighting now works upon label & tree hover and effectively turns off when the cursor leaves a linked view area.
Fixed a minor bug which caused the last cell of a row or column to not
be included in the highlight.

Prevented the hover highlight of rows/cols to activate while a selection
is being made.

I also simplified code & renamed some variables and methods to make it
more intuitive and efficient.

Moved the highlight state into MapContainer so that I could turn off
hover highlighting from IMV, LabelView, and TRView upon a mousePressed
event (to start selecting).

I had to un-deprecate (i.e. remove a 'deprecate' comment) from the
getXMap and getYMap methods in IMV.  I needed them in the
MatrixViewController to know whether or not to start the highlight mode.
It should not start during a selection.  I.e. if the user switches from
a regular selection to a row selection by depressing the shift key
during a drag, the row hover highlight should not turn on.