Replacing confusingly named 'header' variables
Many classes, methods and variables contain 'header' as part of their
name. It is used interchangeably as 'label' and 'label header' or 'label
prefix' which makes code VERY hard to understand.

This commit is the first which replaces this old naming with something
more consistent: prefixes and labels.

Prefixes refer to the name of a label array, such as ORF, ROW_ID etc.
Labels refer to what is the commonly understood term: a label for either
a specific row or column.

The commit contains MOSTLY refactoring according to the pattern
described. Minor other renaming (gene -> row, array --> column) has also
been performed.