Added no-labels check for LabelView.getMaxStrLen()
getMaxStrLen() accesses the LabelSummary object to retrieve labels and
calculate the maximum length of all labels. If no labels have been
loaded or designated, it would still attempt to access them. This led to
ArrayOutOfBoundsExceptions in LabelSummary and IntLabelInfo. These were
also printed to the console ad nauseum, causing the application to slow.