Fixed array access issue in LabelFinderBox
findSelected() had no check for the value of primarySearchIndex and
would attempt to access label type columns using it as an index, even if
the value was -1.

Also changed the name of searchDataList because it isn't a List in the
Java sense. It is an array and has thus been named searchDataArray to
avoid confusion.