Fixed never changing member in getMaxStrLength
The LabelView data member longest_str was never updated by
getMaxStrLength unless debug mode was on, because the local variable
maxStr was never updated.

This caused certain if-clauses to fire at
every repaint of LabelView although the maximum string length has
already been determined. Consequently, the method thought a
recalculation of the maximum length string index was needed and an
infinite loop of label-walking for-loops was started.

The problem still continues because other variables are not properly
updated (lastDrawnFace etc.). This will be addressed next.