Created LabelAttributes class
The label attributes class encapsules the management of label attributes
and takes this responsibility away from LabelView. This makes debugging
a lot easier and offers a central point for handling label attributes.

Refactoring for min and max size methods in font related classes has
also been performed.

This class has NOT been put into use yet and LabelView is in its old
form. This will be addressed next.

The change is a result of complications when debugging getMaxStrLength.
A simpler interface for understanding the state of label attributes is
enabling easier debugging for #447.