DataMatrix reset full reset ensured
The DataMatrix object of the TVModel was not fully reset when
resetState() was called on the TVModel. Min/max/median/mean were all not
reset to their initial default values which could lead to
calculateBaseValues() delivering false results.

The method should be decoupled from the state of the instance variables
aside from exprData so that such errors cannot occur.