Added the ability to change the amount of real estate dedicated to the global view, label splitpanes, and the interactive matrix.  Also fixed a bug that caused the cursor icon to not change at the intersection as well as a bug which caused the intersection cursor to not be correct on macs.
There are issues this creates with the labels and trees syncing up with
the matrix and the drawing of the currend position in the global view
and the selection of tiles.  But as a first step, the functionality
works.

Also, the initial setup of the window gives all the real estate to the
global view.  Haven't figured out how to fix that yet.