Fixed the initial layout.
The setWidths and setHeights methods were incorrectly normalizing the
values so that they would sum to 1.  All I had to do was change the
scope of the for loop which summed the values to not include the
normalization loops.

I also mucked around with the colors of the cell boundaries to try and
get them to go from the background color to white upon hover & drag, but
those changes aren't working correctly yet.