Made it possible to keep static grid cell sizes (if they were initially set statically via the `int setWidths` and `int setHeights` methods) when the window is resized.  Thus the only grid cell that changes dimensions upon window resize is the interactive matrix.
Note, I had to add a check to make sure that resizing a window smaller
neither pushes grid cells completely off the window nor resizes grid
cells smaller than the minimum.

There might be the possibility here for an exception if the window size
is made to be smaller than the sum of the minimum sizes, but currently
that's not an issue because the minimum is much smaller than the
smallest possible window size.