Changed the draggrid border color to light gray to make it easier to see against the matrix and added a call to both the map containers' AdjustToScreenChange inside MatrixView's updateMatrix method.
I had also made the dragGrid object a member of DendroView in a failed
attempt to make it communicate with TVController to let it know when the
dragGrid cells sizes have changed (upon dragging a border) so that
TVController could call the various update methods on the label, tree,
etc objects to adjust to the new sizes.  However, I was unable to figure
out how to establish that communication.  Turned out it was unnecessary.
All I had to do was update the available pixels in the mapcontainers via
the added call mentioned above.  That method in MatrixView already
catches the screen real estate change, which I noted by the fact that
the matrix was always drawn correctly.  However, I left dragGrid as a
member of DendroController because it just seemed to make sense that
way.

If there's a better way to initiate the update of the available pixels
in MapContainer, I'm all ears.