Got the generation of the .app directory to work using gradle.
The app icon is now applied to the .app "file" and the name when you
hover over it in the dock, or see it in the Finder is "TreeView3" of
"TreeView3.app" depending on the context.

Note, I had to trim some border around the 2 logo.png files that exist
in the project to be square.  It was just transparent background I
trimmed.  I also created a .icns file needed by Mac OS X to apply as an
app icon.