Handling data coords detection failure
Automatically detecting data boundaries can lead to failure when no
valid numeric elements can be detected in concert with the given restrictions,
particularly when there are many rows or columns. The default data
coordinates of -1 and Integer.Max were not changed which lead OutOfMemory to errors
when rendering the preview data table cells. A finally block now ensures a
default setting for data start coordinates in case they cannot be
conventionally detected.