Partially working label export.
I'm only committing so that I can pull the current changes on another
computer I'm working on today.

I partially implemented the inclusion of labels in the exported "all"
region (i.e. not the visible or selected regions).  It's buggy.  Labels
are drawn currently, but not in the right place and the space for them
was not created.