Added checks for final image size when labels are included.
Export dialog options are enabled/disabled based on the selected label
options and follow a hierarchy of Format, Region, Aspect, Row labels,
and Column labels.  Items higher in the hierarchy assume the user
selections made and lower assume the minimum dimension options.  Row
labels assume the selected column label option and vice versa.  What
this basically means is, selecting a bottom option does not affect the
enabledness of a higher option.  Selecting a higher option can affect
the enabledness of a lower option.  If a user's lower level selection
becomes invalid when they select a higher option, their selection is
changed to the minimum default in order to be valid.  E.g. If the region
selected in the "selection" and they've selected row and column labels,
then change the region to "all", resulting in the image being too big,
the label options are deselected until the image is small enough to be
exportable.