Fixed a bug in the calculation of the label area length.
In some contexts, the longest label length was being calculated among
the selected labels (if only drawing selected labels). In other
contexts, the length was calculated among all labels. In all contexts,
the exported region was ignored and all labels (or all selected labels)
were inspected.  Now everywhere (WRT export), the region and whether
we're only drawing selected labels are taken into account.  The purpose
of this is so there are no huge gaps between the labels and the trees
(or matrix - depending on set justification).