Fixed a bug where column labels were being drawn too low.
The orientLabelPane method that was being called was using the max label
length in its call to translate.  (Actually, it was the offscreen
height, but that's what it's set to - using the font size in the GUI
currently displayed - not the font size of the export either.)  So I
created an orientForPreview method which additionally takes the max
label length.