Fixed a few bugs I encountered.
The code to update the label radio buttons was not working because the
dimensions weren't calculated to accommodate the labels.  I don't know
how this wasn't working before.  Perhaps I broke it when I started
addressing logic to handle the matrix embedded inside a PDF...

I added logic to disable/enable export options when a document format is
selected, since Faizan's embedding of a PNG in a PDF made it susceptible
to the size limit imposed by BufferedImage.

I found a bug where row should have been column or vice versa.  It was
in the code that enabled/disabled the export buttons.  I think that it
didn't really matter much since if you include either row or column
labels, both dimensions are affected to maintain the aspect ratio.

I added some documentation to some methods because it wasn't immediately
apparent why some code was doing what it did.