This is a step closer to fixing the memory bug WRT labels.
Labels *should* no longer cause a memory error.  I accomplished this by
actually drawing labels down to a font size of 2.  If they are any
smaller than that, lines are drawn instead of labels in an image that is
not taller/wider than the matrix.

So the prevention of the memory bug is dependent on the matrix not being
too big.  If the matrix is too big, you won't be able to export anyway.

I did encounter a memory error about the preview image after
implementing these changes, however it was from selecting "as seen on
screen" for the 1kx1k matrix (which is really 1kx6k) and the actual
export also encounters an out of memory error, so that's not specific to
the labels.