Improved the out of memory error messages and set the export status to unsuccessful so the non-existent exported file will not be attempted to be opened.
There are still exceptions that happen related to a linked list that are
triggered after the failed export stops and they appear to be triggered
by mouse actions while the export dialog is up.  actionPerformed is thus
seems to be getting called by things other than clicking the export
button.