Got the update of the buttons in the export dialog to properly update.
I created getMinDefault methods in the various enumerations that the
export dialog uses to track options (e.g. RegionType) so that defaults
could be set to whatever is desired as a default and if it was too big,
the minimum default could be a fallback.  I used this to turn on label
export by default.

There is still an immediate java heap size error when you try to export
the larger_1kx1k_average.cdt file (which is actually a 1k by 5k file)
with the default options.  The resulting image is not too big according
to the BufferedImage class limit.  Not sure what is happening here, but
this problem is beyond the scope of this branch.