Fixed a bunch of weirdness in the export preview.
I'd forgotten to call setCalculatedDimensions after setting the export
options.  Adding that call fixed just about everything wrong that was
happening.  The only remaining niggling issue is that sometimes some
tree branches don't show up in the preview.  This has to do with
resizing, but is not specific to this branch/PR.