Fully resolved issue 9 from pull request #133 - Export situations that have no possible valid export given size now prevent the export dialog from popping up.
Also, if a document format export is possible, but image export is not,
the format select list is adjusted to not provide image export options.
Note, it was not possible using JComboBox to disable the image options,
but still leave them visible.

Switching formats also triggers updates of the other dialog elements.