The functionality for only drawing a certain number of labels flanking the hovered position is implemented.
There is no way yet for the user to change the behavior.  The default
behavior is set to 0 flanking labels.

If flanking label mode is on and the flank size is 0, the indicator bar
is not drawn and the label color is kept black.

Neither labels nor selected background color are drawn in flanking mode
if the labels are greater than the flak size away from the hover
position.

All possible settings for the flanking size have been tested: less than
0, 0, a value greater than 0 but less than the labels that can fit, and
a value greater than both 0 and the number that fit in the label pane.