I implemented spacebar toggling of the way labels are displayed.
It works, but there's a minor bug where the indent of the labels is not
immediately updated when switching between the two modes.  This has been
an issue for awhile actually in other contexts, but happens every time
in this context, particularly WRT the column labels.  It likely has to
do with which way the justification is.