Swaped the modifier for animated zoom with the default button click behavior for the zoom selection and home buttons.
Clicking either button without a modifier now zooms without animation.
Holding down the command key zooms with animation.

This is admittedly unrelated to issue 533, but is overall related to the
intent behind 533.  It was all discussed as 1 issue in our last meeting.

I basically copied the code changes in the tv3_no_animations branch.