Resolved issue 534 - label settings not remembered.
The state was being stored in the various setters for each
font-associated setting, which was overwriting the saved settings
whenever the defaults were loaded upon startup.

Also, needed to make one more call to a synchronize method and had to
correct the storage of the fixed & justified settings, which were always
saving the default instead of the user-supplied settings.