Improved an incorrect assumption that if either GID is present in the CDT file OR the file exists, that the file is clustered.
Now a file is only considered to be clustered if BOTH are true.

Also added a note to an error message about a file having been clustered
before so that it refers to where you can find other clustered versions
of the file.

Added a try/catch around mismatching GIDs/AIDs so that clustering
doesn't completely fail.