Fixed the drawing of the trees when drawing fewer than all labels that can fit the label area.
Apparently, I completely overlooked and did not test the trees when I
implemented the new label features.  Here's the new behavior: 1. When
only 1 label is being drawn, the tree remains visually associated with
the data in the matrix.  2. When more the 1 label is drawn (but fewer
than can actually fit) and the cursor is not over the labels, again -
the tree remains visually linked to the data and does not move, but the
blue bacground behind the tree expands to include all the labels that
are being drawn.  3. In the same circumstance as #2, but when the cursor
is over the labels, the tree is visually linked to the labels and moves
with them.  No blue background is drawn.