Addressed label flicker upon held spacebar issue.
Note that this means that if the user hits the spacebar in rapid
succession, the second press will not toggle because the software
assumes that rapidly repeating spacebar presses are the result of
holding the spacebar down.  The toggle will not revert when a held
spacebar is released.  Note that I tried implementing this, but there's
always at least a noticeable 50 millisecond delay for the display to
revert.  There's no other way to prevent the flicker AND implement the
behavior where labels are only shown during the holding down of the
spacebar.  The flicker can only be prevented via a timer, thus there
would always be a delayed response upon release.  Given the toggle
implkementation (as we have, the response is immediate and holding the
spacebar means nothing).  The timer simply prevents the flickering and
the side effect is that rapid intentional pressings of the spacebar will
not always work on the subsequent pressings.