Resolved issue 559 - GMV location wrong.
I just needed to update the matrix once it was done loading.  There's a
minimum size for the viewport rectangle and when either setXMap or
setYMap are called, the sizes there are not initialized to the newly
loaded data.  In fact, I should remove that useless code for setting the
minimum viewport size in those methods for that reason.