Removed useless setXMap and setYMap methods.
The xmap and ymap objects still have old data in them when these methods
were called, thus the xViewMin and yViewMin were getting improperly set.
The code was copied to updateMatrix in the previous commit (at which
point, the objects are properly populated).

I also added javadoc for that method.