This resolves issue 547 - blank load on Ubuntu for large data files.
There were 2 issues:

1. An out of memory exception was silently occurring in a SwingWorker
thread and the main thread proceeded without realizing that the data was
not fully loaded.

2. 2 copies of the large data matrix were being created, which was
unnecessary.  While it may be unnecessary to create any copies, I
eliminated the need for the second copy by retrieving the array index of
the first occurrence of a NaN or infinite value and then using that
value to get the median from the un-truncated array.

I also noted that when the failure occurs, the user is "returned" to the
openFile dialog, even if they used the "Open Last File" button.  I will
create an issue for that.