Added a catch to the exception thrown from the SwingWorker thread and set the message given to the user to have a tidy message in the instance of an out of memory exception.
I tested this by adding a line of code to copy the array while
calculating the median again to cause the out of memory exception on the
Ubuntu VM I was using for testing.

The exception that's caught is actually an ExecutionException wrapping
the out of memory exception.

I also removed a bunch of debug outputs and cleaned up some code.