This version of build.gradle actually somewhat works - for creating a working .deb file for distribution.
It puts the jar and a shell script in /opt/local/treeview/{lib,bin} respectively.

The shell script needs to be tested.  An icon is still needed.  Vestigial code in the gradle script needs to be weeded out.  A link from a bin in the user's path needs to be created.