This version puts things in the right place and actually works.
The last checkin was putting things in the right place, but the shell script was failing because it wasn't including the fat jar in the classpath argument to java.  I just gave up on the fat jar and copied all the jar files to where they are expected using nebula/ospackage.

I took out some vestigial code.