Added a code signing certificate for Mac builds so that users do not get the "Unidentified developer" error.
I included a document in the docs/softeng folder describing how to
generate & install a certificate so that gradle can perform the
code-signing.

Note, users will still get a warning that the app was downloaded from
the internet (though it doesn't say it's an unknown developer).  The
only way to get around that is to distribute the app in the Apple App
Store.