Changed the windows batch script to check the exit status of the attempt to start TreeView and take other actions.
First it tries to use the JAVA_HOME variable and decreasing amounts of
memory.  Then it assumes that java is in the path and again tries to
start with 4G memory and decreases from there.  Finally, if none of
those worked (meaning they probably don't have java installed), it
throws up a command prompt window with an error message echoed to the
screen.